+++ app/src/main/java/com/example/videoapp/di/ApiModule.kt
@@ -66,7 +66,6 @@ object ApiModule {
             .build()
     }
 
-
     @Provides
     @Singleton
     fun provideRetrofit(baseUrl: String, gson: Gson, client: OkHttpClient): ApiServices =
