+++ app/src/main/java/com/example/videoapp/paging/MovieCategoryPagingSource.kt
@@ -28,9 +28,7 @@ class MovieCategoryPagingSource(
         } catch (exception: HttpException) {
             LoadResult.Error(exception)
         }
-
     }
-
 
     override fun getRefreshKey(state: PagingState<Int, MoviesListResponse.Result>): Int? {
         return null
