+++ app/src/main/java/com/example/videoapp/ui/CategoryFragment.kt
@@ -4,13 +4,12 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
-import android.widget.ArrayAdapter
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
+import androidx.navigation.fragment.navArgs
 import androidx.paging.LoadState
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.videoapp.adapter.LoadMoreAdapter
@@ -30,9 +29,7 @@ class CategoryFragment : Fragment() {
 
     private val viewModel: MoviesViewModel by viewModels()
 
-    private val category = arrayOf("Action", "History", "Adventure","Horror","Animation","Music","Comedy"
-        ,"Mystery","Crime","Romance","Documentary","Science","Drama","TV Movie","Family","Thriller"
-        ,"Fantasy","War","Western")
+    private val args: CategoryFragmentArgs by navArgs()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -42,104 +39,45 @@ class CategoryFragment : Fragment() {
         return binding.root
     }
 
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val spinner = binding.categorySpinner
-        ArrayAdapter(
-            requireContext(),
-            android.R.layout.simple_spinner_item
-            ,category
-        ).also { adapter ->
-            adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
-            spinner.adapter = adapter
-        }
 
-        binding.categorySpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener{
-            override fun onItemSelected(
-                parent: AdapterView<*>?,
-                view: View?,
-                position: Int,
-                id: Long
-            ) {
-                val categoryNumber: String = convertCategoryToNumber(category[position])
+        val categoryNumber: String = convertCategoryToNumber(args.movieCategory)
 
-                binding.apply {
-
-                    lifecycleScope.launchWhenCreated {
-                        viewModel.moviesCategoryList(categoryNumber).collect {
-                            moviesAdapter.submitData(it)
-                        }
-                    }
-
-                    moviesAdapter.setOnItemClickListener {
-                        val direction = MoviesFragmentDirections.actionMoviesFragmentToMovieDetailsFragment(it.id)
-                        findNavController().navigate(direction)
-                    }
-
-                    lifecycleScope.launchWhenCreated {
-                        moviesAdapter.loadStateFlow.collect{
-                            val state = it.refresh
-                            prgBarMovies.isVisible = state is LoadState.Loading
-                        }
-                    }
-
-
-                    rlMovies.apply {
-                        layoutManager = LinearLayoutManager(requireContext())
-                        adapter = moviesAdapter
-                    }
-
-                    rlMovies.adapter=moviesAdapter.withLoadStateFooter(
-                        LoadMoreAdapter{
-                            moviesAdapter.retry()
-                        }
-                    )
-
+        binding.apply {
+            lifecycleScope.launchWhenCreated {
+                viewModel.moviesCategoryList(categoryNumber).collect {
+                    moviesAdapter.submitData(it)
                 }
             }
 
-            override fun onNothingSelected(parent: AdapterView<*>?) {
-                binding.apply {
+            moviesAdapter.setOnItemClickListener {
+                val direction = CategoryFragmentDirections
+                    .actionCategoryFragmentToMovieDetailsFragment(it.id)
+                findNavController().navigate(direction)
+            }
 
-                    lifecycleScope.launchWhenCreated {
-                        viewModel.moviesList.collect {
-                            moviesAdapter.submitData(it)
-                        }
-                    }
-
-                    moviesAdapter.setOnItemClickListener {
-                        val direction = MoviesFragmentDirections.actionMoviesFragmentToMovieDetailsFragment(it.id)
-                        findNavController().navigate(direction)
-                    }
-
-                    lifecycleScope.launchWhenCreated {
-                        moviesAdapter.loadStateFlow.collect{
-                            val state = it.refresh
-                            prgBarMovies.isVisible = state is LoadState.Loading
-                        }
-                    }
-
-
-                    rlMovies.apply {
-                        layoutManager = LinearLayoutManager(requireContext())
-                        adapter = moviesAdapter
-                    }
-
-                    rlMovies.adapter=moviesAdapter.withLoadStateFooter(
-                        LoadMoreAdapter{
-                            moviesAdapter.retry()
-                        }
-                    )
-
+            lifecycleScope.launchWhenCreated {
+                moviesAdapter.loadStateFlow.collect{
+                    val state = it.refresh
+                    prgBarMovies.isVisible = state is LoadState.Loading
                 }
             }
 
-        }
+            rlMovies.apply {
+                layoutManager = LinearLayoutManager(requireContext())
+                adapter = moviesAdapter
+            }
 
+            rlMovies.adapter=moviesAdapter.withLoadStateFooter(
+                LoadMoreAdapter{
+                    moviesAdapter.retry()
+                }
+            )
+        }
     }
 
-    fun convertCategoryToNumber(category: String): String {
+    private fun convertCategoryToNumber(category: String): String {
         when(category){
             "Action" -> return "28"
             "Adventure" -> return "12"
