+++ app/src/main/java/com/example/videoapp/ui/MainActivity.kt
@@ -13,13 +13,14 @@ import dagger.hilt.android.AndroidEntryPoint
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
+
     private lateinit var navController: NavController
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         navController = Navigation.findNavController(this,R.id.navHost)
-        setupWithNavController(binding.bottomNavigationView,navController)
     }
 }
