+++ app/src/main/java/com/example/videoapp/ui/MoviesFragment.kt
@@ -1,9 +1,12 @@
 package com.example.videoapp.ui
 
+import android.R
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
@@ -28,6 +31,9 @@ class MoviesFragment : Fragment() {
 
     private val viewModel: MoviesViewModel by viewModels()
 
+    private val category = arrayOf("Category name","Action", "History", "Adventure","Horror","Animation","Music","Comedy"
+        ,"Mystery","Crime","Romance","Documentary","Science","Drama","TV Movie","Family","Thriller"
+        ,"Fantasy","War","Western")
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -39,6 +45,17 @@ class MoviesFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        val spinner = binding.categorySpinner
+        ArrayAdapter(
+            requireContext(),
+            R.layout.simple_spinner_item
+            ,category
+        ).also { adapter ->
+            adapter.setDropDownViewResource(R.layout.simple_spinner_dropdown_item)
+            spinner.adapter = adapter
+        }
+
         binding.apply {
 
             lifecycleScope.launchWhenCreated {
@@ -59,7 +76,6 @@ class MoviesFragment : Fragment() {
                 }
             }
 
-
             rlMovies.apply {
                 layoutManager = LinearLayoutManager(requireContext())
                 adapter = moviesAdapter
@@ -70,7 +86,26 @@ class MoviesFragment : Fragment() {
                     moviesAdapter.retry()
                 }
             )
+        }
 
+        binding.categorySpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener{
+            override fun onItemSelected(
+                parent: AdapterView<*>?,
+                view: View?,
+                position: Int,
+                id: Long
+            ) {
+                if (position > 0) {
+                    val direction =
+                        MoviesFragmentDirections
+                            .actionMoviesFragmentToCategoryFragment(category[position])
+                    findNavController().navigate(direction)
+                    binding.categorySpinner.setSelection(0)
+                }
+            }
+
+            override fun onNothingSelected(parent: AdapterView<*>?) {
+            }
         }
 
         binding.btnSearch.setOnClickListener {
@@ -78,7 +113,5 @@ class MoviesFragment : Fragment() {
             val direction = MoviesFragmentDirections.actionMoviesFragmentToMoviesSearchFragment(searchMovieName)
             findNavController().navigate(direction)
         }
-
     }
-
 }
