+++ app/src/main/res/layout/fragment_movies.xml
@@ -7,12 +7,28 @@
     android:background="#AEBDCA"
     tools:context=".ui.MoviesFragment">
 
+    <Spinner
+        android:id="@+id/categorySpinner"
+        android:layout_width="380sp"
+        android:layout_height="50sp"
+        android:layout_marginStart="10sp"
+        android:layout_marginTop="5sp"
+        android:layout_marginEnd="10sp"
+        android:autofillHints="category"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.476"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.014"
+        tools:ignore="SpeakableTextPresentCheck" />
+
     <EditText
         android:id="@+id/movieName"
-        android:layout_width="271dp"
-        android:layout_height="48dp"
+        android:layout_width="270sp"
+        android:layout_height="50sp"
         android:layout_marginStart="10dp"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="80sp"
         android:layout_marginBottom="22dp"
         android:autofillHints=""
         android:ems="10"
@@ -28,8 +44,8 @@
 
     <Button
         android:id="@+id/btnSearch"
-        android:layout_width="107dp"
-        android:layout_height="48dp"
+        android:layout_width="100sp"
+        android:layout_height="50sp"
         android:layout_marginStart="10dp"
         android:layout_marginEnd="10dp"
         android:text="@string/search"
@@ -41,14 +57,15 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rlMovies"
         android:layout_width="match_parent"
-        android:layout_height="720sp"
-        android:layout_marginTop="100dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="150sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
+
         app:layout_constraintVertical_bias="0.0" />
 
     <ProgressBar
