--- app/src/main/res/menu/nav_menu.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/moviesFragment"
-        android:icon="@drawable/baseline_home_24"
-
-        android:title="Home"
-        />
-    <item
-        android:id="@+id/categoryFragment2"
-        android:icon="@drawable/baseline_category_24"
-        android:title="Category"
-        />
-
-</menu>
