+++ app/src/main/res/navigation/nav_main.xml
@@ -16,6 +16,9 @@
         <action
             android:id="@+id/action_moviesFragment_to_moviesSearchFragment"
             app:destination="@id/moviesSearchFragment" />
+        <action
+            android:id="@+id/action_moviesFragment_to_categoryFragment"
+            app:destination="@id/categoryFragment" />
     </fragment>
     <fragment
         android:id="@+id/movieDetailsFragment"
@@ -38,8 +41,15 @@
             app:destination="@id/movieDetailsFragment" />
     </fragment>
     <fragment
-        android:id="@+id/categoryFragment2"
+        android:id="@+id/categoryFragment"
         android:name="com.example.videoapp.ui.CategoryFragment"
         android:label="fragment_category"
-        tools:layout="@layout/fragment_category" />
+        tools:layout="@layout/fragment_category" >
+        <action
+            android:id="@+id/action_categoryFragment_to_movieDetailsFragment"
+            app:destination="@id/movieDetailsFragment" />
+        <argument
+            android:name="movieCategory"
+            app:argType="string" />
+    </fragment>
 </navigation>
